refactor: address -> map으로 패키치명 수정
